oxi plhrws leitourgiko, tha to ftiaksw meta

+++ reasy_android_studio/Reasy/app/src/main/java/com/example/reasy/catering.java
@@ -5,9 +5,9 @@ public class catering {
     private String name;
     private int catering_id;
     private String cuisine_type;
-    private float cost;
+    private double cost;
 
-    public catering(String name, int catering_id, String cuisine_type, float cost) {
+    public catering(String name, int catering_id, String cuisine_type, double cost) {
         this.name = name;
         this.catering_id = catering_id;
         this.cuisine_type = cuisine_type;
