+++ reasy_android_studio/Reasy/app/src/main/java/com/example/reasy/job_offer.java
@@ -4,14 +4,14 @@ public class job_offer {
 
     private int shop_id;
     private String position;
-    private float salary;
+    private double salary;
     private ArrayList<String> working_hours = new ArrayList<String>();
-    private float experience;
+    private double experience;
     private String start_date;
     private String end_date;
     private String description;
 
-    public job_offer(int shop_id, String position, float salary, ArrayList<String> working_hours, float experience, String start_date, String end_date, String description) {
+    public job_offer(int shop_id, String position, double salary, ArrayList<String> working_hours, double experience, String start_date, String end_date, String description) {
         this.shop_id = shop_id;
         this.position = position;
         this.salary = salary;
@@ -21,7 +21,7 @@ public class job_offer {
         this.end_date = end_date;
         this.description = description;
     }
-    public float getSalary(){
+    public double getSalary(){
         return salary;
     }
 }
