+++ reasy_android_studio/Reasy/app/src/main/java/com/example/reasy/menu.java
@@ -5,9 +5,9 @@ import java.util.ArrayList;
 public class menu {
     private int shop_id;
     private ArrayList<product_menu> products=new ArrayList<product_menu>();
-    private float rating;
+    private double rating;
     private int num_of_rates;
-    public menu(int shop_id, ArrayList<product_menu> products, float rating, int num_of_rates) {
+    public menu(int shop_id, ArrayList<product_menu> products, double rating, int num_of_rates) {
         this.shop_id = shop_id;
         this.products = products;
         this.rating = rating;
@@ -17,8 +17,8 @@ public class menu {
     public ArrayList<product_menu> getProducts(){
         return products;
     }
-    public void changeRatingMenu(float new_r){
-        float r=num_of_rates*rating;
+    public void changeRatingMenu(double new_r){
+        double r=num_of_rates*rating;
         num_of_rates++;
         rating=(rating+new_r)/num_of_rates;
     }
