+++ reasy_android_studio/Reasy/app/src/main/java/com/example/reasy/rating.java
@@ -6,9 +6,9 @@ public class rating {
 
     private int customer_id;
 
-    private int evaluation;
+    private double evaluation;
 
-    public rating(int shop_id, int customer_id, int evaluation) {
+    public rating(int shop_id, int customer_id, double evaluation) {
         this.shop_id = shop_id;
         this.customer_id = customer_id;
         this.evaluation = evaluation;
