+++ reasy_android_studio/Reasy/app/src/main/java/com/example/reasy/supply.java
@@ -4,7 +4,7 @@ import java.util.ArrayList;
 
 public class supply {
 
-    public supply(int supply_id, int shop_id, int supplier_id, ArrayList<product_supply> products, String address, boolean sample, float cost) {
+    public supply(int supply_id, int shop_id, int supplier_id, ArrayList<product_supply> products, String address, boolean sample, double cost) {
         this.supply_id = supply_id;
         this.shop_id = shop_id;
         this.supplier_id = supplier_id;
@@ -20,5 +20,5 @@ public class supply {
     private ArrayList<product_supply> products = new ArrayList<product_supply>();
     private String address;
     private boolean sample;
-    private float cost;
+    private double cost;
 }
